requirements.txt
streamlit
pandas
google-search-results
google-generativeai